refactor: Moved tests
